# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: C5 360相机巡查标注系统 API
  version: 1.0.0
  license:
    name: Proprietary
  description: >
    Spec-first OpenAPI 3 contract for project C5 (360-camera inspection &
    annotation system). This single document is the authoritative source for
    oapi-codegen v2 → Go server stubs + TypeScript + Kotlin/Retrofit clients.


    Conventions:

    - Every response uses the shared envelope `{success, data, error, meta?}`.
      List endpoints populate `meta` with `{total, page, page_size}`.

    - All IDs are int64 (BIGSERIAL in the DB). `client_uuid` / `job_uuid` are
      UUID strings (offline idempotency / public job handles).

    - All geometry on the wire is WGS84 / SRID 4326 GeoJSON. GCJ-02 is NEVER
      persisted or transmitted; it is a client-only render-boundary concern.

    - All timestamps are RFC3339 / ISO-8601 UTC (`format: date-time`).

    - Error codes are a frozen, canonical enum (see ErrorObject.code).


    Out-of-band (NOT expressible in OpenAPI):

    - `GET /exports/{job_uuid}/events` is a Server-Sent Events (`text/event-stream`)
      endpoint emitting `event: progress` and a terminal `event: done`. It is
      hand-written with `GET /exports/{job_uuid}` as the polling fallback
      (Tencent CLB/CDN may buffer SSE). It is intentionally absent from `paths`.

    - Dictionary / config pull uses ETag (`If-None-Match: "<content_hash>"` →
      304 Not Modified when unchanged). 304 responses carry no body.

...
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /problems:
    get:
      tags: [problems]
      operationId: listProblems
      summary: List problems (D1 filters)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: project_id
          in: query
          schema: { type: integer, format: int64 }
        - name: type
          in: query
          description: type_item_id (dict_item id, scope problem_type)
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          description: status_item_id (dict_item id, scope problem_status)
          schema: { type: integer, format: int64 }
        - name: category
          in: query
          description: category_item_id (dict_item id, scope problem_category)
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/FromParam'
        - $ref: '#/components/parameters/ToParam'
        - name: inspector_id
          in: query
          schema: { type: integer, format: int64 }
        - name: inspection_id
          in: query
          description: D1 — filter problems by inspection session
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Problem list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemListResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [problems]
      operationId: createProblem
      summary: Create a problem annotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProblemCreate'
      responses:
        '201':
          description: Problem created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
  /problems/map:
    get:
      tags: [problems]
      operationId: getProblemsMap
      summary: Problems as a GeoJSON FeatureCollection (WGS84) for the map layer
      parameters:
        - name: project_id
          in: query
          schema: { type: integer, format: int64 }
        - name: type
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema: { type: integer, format: int64 }
        - name: category
          in: query
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/FromParam'
        - $ref: '#/components/parameters/ToParam'
        - name: inspector_id
          in: query
          schema: { type: integer, format: int64 }
        - name: inspection_id
          in: query
          schema: { type: integer, format: int64 }
        - name: limit
          in: query
          description: Server-enforced cap to avoid full-table scans
          schema: { type: integer, default: 5000, maximum: 20000 }
      responses:
        '200':
          description: GeoJSON FeatureCollection of problem points
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemMapResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /problems/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [problems]
      operationId: getProblem
      summary: Get a problem by id
      responses:
        '200':
          description: Problem
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [problems]
      operationId: updateProblem
      summary: Update a problem
      description: >
        D3 — if `status_item_id` changes, the backend atomically appends a
        `STATUS_CHANGE` row to `problem_processing_log` in the SAME transaction.
        Clients never POST `STATUS_CHANGE` directly.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProblemUpdate'
      responses:
        '200':
          description: Problem updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [problems]
      operationId: deleteProblem
      summary: Soft-delete a problem
      responses:
        '200':
          description: Problem deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /problems/{id}/logs:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [problems]
      operationId: listProblemLogs
      summary: List a problem's processing logs (append-only audit trail)
      responses:
        '200':
          description: Processing log list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemProcessingLogListResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [problems]
      operationId: createProblemLog
      summary: Append a COMMENT or REASSIGN log (D3 — STATUS_CHANGE is backend-only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProblemLogCreate'
      responses:
        '201':
          description: Log appended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemProcessingLogResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # Media
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access JWT (golang-jwt/v5). Refresh is an opaque Redis token.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
    FromParam:
      name: from
      in: query
      description: Inclusive lower time bound (RFC3339 UTC)
      schema: { type: string, format: date-time }
    ToParam:
      name: to
      in: query
      description: Exclusive upper time bound (RFC3339 UTC)
      schema: { type: string, format: date-time }
...
  responses:
...
    Unauthenticated:
      description: Missing/invalid/expired bearer token (UNAUTHENTICATED | TOKEN_EXPIRED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: casbin (sub,obj,act) denied (FORBIDDEN)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found or soft-deleted (NOT_FOUND)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Unique conflict / RESTRICT / illegal state transition (CONFLICT)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationFailed:
      description: Boundary schema / business validation failed (VALIDATION_FAILED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Unhandled error (INTERNAL); never leaks stack/SQL
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Meta:
      type: object
      properties:
        total: { type: integer, format: int64 }
        page: { type: integer }
        page_size: { type: integer }
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - UNAUTHENTICATED
        - TOKEN_EXPIRED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - MEDIA_VERIFY_FAILED
        - DICT_VERSION_RETIRED
        - RATE_LIMITED
        - INTERNAL
    ErrorDetail:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message: { type: string }
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: false }
        data:
          type: object
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    EmptyResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'

    # -------------------------------------------------------------------------
    # GeoJSON primitives (WGS84 / 4326)
    # -------------------------------------------------------------------------
    GeoJSONPoint:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [Point]
        coordinates:
          type: array
          description: '[longitude, latitude] in WGS84'
          items: { type: number }
          minItems: 2
          maxItems: 2
...
    ProblemFeatureProperties:
      type: object
      required: [id]
      properties:
        id: { type: integer, format: int64 }
        type_item_id: { type: integer, format: int64, nullable: true }
        status_item_id: { type: integer, format: int64, nullable: true }
        category_item_id: { type: integer, format: int64, nullable: true }
        type_label: { type: string, nullable: true }
        status_label: { type: string, nullable: true }
        status_color: { type: string, nullable: true }
        title: { type: string }
        captured_at: { type: string, format: date-time }
        cover_media_id: { type: integer, format: int64, nullable: true }
    ProblemFeature:
      type: object
      required: [type, geometry, properties]
      properties:
        type:
          type: string
          enum: [Feature]
        geometry:
          $ref: '#/components/schemas/GeoJSONPoint'
        properties:
          $ref: '#/components/schemas/ProblemFeatureProperties'
    ProblemFeatureCollection:
      type: object
      required: [type, features]
      properties:
        type:
          type: string
          enum: [FeatureCollection]
        features:
          type: array
          items:
            $ref: '#/components/schemas/ProblemFeature'

    # -------------------------------------------------------------------------
    # Native enums (exact canonical literals)
    # -------------------------------------------------------------------------
...
    ProcessingAction:
      type: string
      description: STATUS_CHANGE is backend-only (D3); clients POST only COMMENT/REASSIGN.
      enum: [STATUS_CHANGE, COMMENT, REASSIGN]
    ClientLogAction:
      type: string
      description: Client-postable subset of ProcessingAction (D3).
      enum: [COMMENT, REASSIGN]
...
    Problem:
      type: object
      required:
        - id
        - client_uuid
        - project_id
        - inspector_id
        - geom
        - dict_version_used
        - captured_at
        - created_at
        - updated_at
      properties:
        id: { type: integer, format: int64 }
        client_uuid: { type: string, format: uuid }
        project_id: { type: integer, format: int64 }
        inspection_id: { type: integer, format: int64, nullable: true }
        inspector_id: { type: integer, format: int64 }
        geom:
          $ref: '#/components/schemas/GeoJSONPoint'
        type_item_id: { type: integer, format: int64, nullable: true }
        status_item_id: { type: integer, format: int64, nullable: true }
        category_item_id: { type: integer, format: int64, nullable: true }
        dict_version_used: { type: integer }
        title: { type: string }
        description: { type: string }
        note: { type: string }
        captured_at: { type: string, format: date-time }
        cover_media_id: { type: integer, format: int64, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProblemProcessingLog:
      type: object
      required: [id, problem_id, action, created_at]
      properties:
        id: { type: integer, format: int64 }
        problem_id: { type: integer, format: int64 }
        action:
          $ref: '#/components/schemas/ProcessingAction'
        from_status_item_id: { type: integer, format: int64, nullable: true }
        to_status_item_id: { type: integer, format: int64, nullable: true }
        note: { type: string }
        operator_id: { type: integer, format: int64, nullable: true }
        created_at: { type: string, format: date-time }
...
    ProblemCreate:
      type: object
      required: [project_id, geom, captured_at]
      properties:
        client_uuid:
          type: string
          format: uuid
          description: Offline idempotency key (server generates if omitted).
        project_id: { type: integer, format: int64 }
        inspection_id: { type: integer, format: int64, nullable: true }
        inspector_id:
          type: integer
          format: int64
          nullable: true
          description: Defaults to the authenticated user if omitted.
        geom:
          $ref: '#/components/schemas/GeoJSONPoint'
        type_item_id: { type: integer, format: int64, nullable: true }
        status_item_id: { type: integer, format: int64, nullable: true }
        category_item_id: { type: integer, format: int64, nullable: true }
        dict_version_used: { type: integer }
        title: { type: string }
        description: { type: string }
        note: { type: string }
        captured_at: { type: string, format: date-time }
        cover_media_id: { type: integer, format: int64, nullable: true }
    ProblemUpdate:
      type: object
      description: >
        D3 — changing status_item_id makes the backend append a STATUS_CHANGE
        log row in the same transaction.
      properties:
        inspection_id: { type: integer, format: int64, nullable: true }
        type_item_id: { type: integer, format: int64, nullable: true }
        status_item_id: { type: integer, format: int64, nullable: true }
        category_item_id: { type: integer, format: int64, nullable: true }
        title: { type: string }
        description: { type: string }
        note: { type: string }
        cover_media_id: { type: integer, format: int64, nullable: true }
    ProblemLogCreate:
      type: object
      required: [action]
      description: D3 — only COMMENT or REASSIGN; STATUS_CHANGE is backend-only.
      properties:
        action:
          $ref: '#/components/schemas/ClientLogAction'
        note: { type: string }
        operator_id:
          type: integer
          format: int64
          nullable: true
          description: For REASSIGN, the new operator/assignee user id.
...
    ProblemResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/Problem'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    ProblemListResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            $ref: '#/components/schemas/Problem'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    ProblemMapResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/ProblemFeatureCollection'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    ProblemProcessingLogResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/ProblemProcessingLog'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    ProblemProcessingLogListResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProblemProcessingLog'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
...